Added testing.xml changes
