+++ component/src/test/resources/testing1.xml
@@ -21,8 +21,8 @@ under the License.
 <suite name="Siddhi-Io-googlepubsub-Test-Suite">
     <test name="Siddhi-io-googlepubsub-tests" enabled="true" preserve-order="true" parallel="false">
         <classes>
-            <class name="org.wso2.extension.siddhi.io.googlepubsub.sink.TestCaseOfGooglePubSubSink"/>
-            <class name="org.wso2.extension.siddhi.io.googlepubsub.source.TestCaseOfGooglePubSubSource"/>
+            <class name="org.wso2.extension.siddhi.io.googlepubsub.sink.TestCaseOfGooglePubSubSink1"/>
+            <class name="org.wso2.extension.siddhi.io.googlepubsub.source.TestCaseOfGooglePubSubSource1"/>
         </classes>
     </test>
 </suite>
